feat: refactor security move itens to auth service
