+++ app/services/auth_service.py
@@ -1,0 +1,57 @@
+from sqlalchemy.orm import Session
+from fastapi import Depends, HTTPException, status
+from fastapi.security import OAuth2PasswordBearer
+from jose import jwt, JWTError
+from app.core.config import settings
+from app.schemas.auth import TokenData
+from app.services.user_service import get_user_by_email
+
+
+from datetime import datetime, timedelta, timezone
+from app.db.session import get_db
+
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
+
+
+def create_access_token(data: dict, expires_delta: timedelta = None):
+    to_encode = data.copy()
+    if expires_delta:
+        expire = datetime.now(timezone.utc) + expires_delta
+    else:
+        expire = datetime.now(timezone.utc) + timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
+    to_encode.update({"exp": expire})
+    encoded_jwt = jwt.encode(to_encode, settings.SECRET_KEY, algorithm=settings.ALGORITHM)
+    return encoded_jwt
+
+
+def decode_token(token: str):
+    credentials_exception = HTTPException(
+        status_code=status.HTTP_401_UNAUTHORIZED,
+        detail="Could not validate credentials",
+        headers={"WWW-Authenticate": "Bearer"},
+    )
+    try:
+        payload = jwt.decode(token, settings.SECRET_KEY, algorithms=[settings.ALGORITHM])
+        email: str = payload.get("sub")
+        scopes: list[str] = payload.get("scopes", [])
+
+        if email is None:
+            raise credentials_exception
+        token_data = TokenData(email=email, scopes=scopes)
+        return token_data
+    except JWTError:
+        raise credentials_exception
+
+
+def verify_token(token: str = Depends(oauth2_scheme), db: Session = Depends(get_db)):
+    token_data: TokenData = decode_token(token)
+
+    user = get_user_by_email(db, email=token_data.email)
+    if user is None:
+        credentials_exception = HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Could not validate credentials",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
+        raise credentials_exception
+    return user
